refactor: Extract table CLI tests to their own module
